hubble/metrics: Introduce v2 HTTP metrics
v2 http metrics are based on L7 response flows, and source/destination
labels are always from the perspective of the request, rather than the
flow. This is consistent with how http metrics are exported in the wider
ecosystem.

By being based on the response flow, we always have the HTTP status
code, making the need for a separate request/response metrics
unnecessary.

In addition, the http_request_duration_seconds metrics now correctly has
the source/destination labels from the perspective of the request,
rather than the reply.

Signed-off-by: Chance Zibolski <chance.zibolski@gmail.com>